Creating and ajusting the entitiy photos to make there the endpoints of photos
