+++ src/resources/users/users.controller.js
@@ -1,7 +1,4 @@
 const userModel = require('./users.model');
-const cloudinary = require('cloudinary').v2;
-const multer = require('multer');
-const upload = multer();
 
 //create user
 const create = async (req, res) => {
@@ -57,74 +54,7 @@ const getUsersByPreferences = async (req, res) => {
   return res.status(200).json(matchedUsers);
 };
 
-const uploadPhoto = async (req, res) => {
-  const imgToUpload = await userModel.createImage({
-    photo: req.file.buffer,
-    user: req.params.id,
-    name: req.file.name,
-    size: req.file.size,
-    mimetype: req.file.mimetype,
-  });
 
-  const photos = await userModel.getImages(req.params.id);
-
-  // Iterate over each image to convert the buffer array into a base64 string
-  all = photos.map((photo) => {
-    return {
-      image: photo.photo.toString('base64'),
-    };
-  });
-  console.log(all);
-
-  return res.status(200).json(all);
-};
-
-//updated photos
-/* const updatePhotos = async (req, res) => {
-    const imagesToUpload = req.files.map(image => {
-        return {
-            image: image.buffer,
-            mimetype: image.mimetype
-        }
-    })
-    // Save all the Images in the array with insertMany()
-    userModel.Photo.updateOne({ 'photoUpload': imagesToUpload  }, (err, imagesToUpload) => {
-        if (err) {
-            res.status(400).send(err);
-        } else {
-            res.status(200).send(imagesToUpload);
-        }
-    }); */
-
-const getByUser = async (req, res) => {
-  const photos = await userModel.getImages(req.params.id);
-  if (photos) {
-    console.log(photos);
-    return res.status(200).json(photos);
-  }
-  return res.status(404).end();
-};
-
-/*  const cloudinary = require('cloudinary').v2;
-cloudinary.config({ 
-    cloud_name: "dcxuyxqvo", 
-    api_key: '387288113589199', 
-    api_secret: '7FCU8VfE3k6IB2oyHxdtts5K9Bo' 
-  });
-
-  const responses = [];
-
-  req.files.forEach(async file => {
-      try {
-          console.log(`data:${file.mimetype};base64,` + file.buffer.toString('base64'))
-        const result = await cloudinary.uploader.unsigned_upload(`data:${file.mimetype};base64,` + file.buffer.toString('base64'), '', '') ;
-      } catch (e) {
-          console.log(e.text)
-      }
-    responses.push(result);
-  })
-
-  console.log(responses); */
 
 module.exports = {
   create,
@@ -133,7 +63,5 @@ module.exports = {
   getOne,
   remove,
   getUsersByPreferences,
-  uploadPhoto,
-  getByUser,
   // getRandom,
 };
