+++ src/resources/users/users.router.js
@@ -14,10 +14,5 @@ router
   .put(usersController.update)
   .delete(usersController.remove);
 
-router
-    .route('/:id/photos')
-    .put(upload.single("photo"), usersController.uploadPhoto)
-    .get(usersController.getByUser);
-
 
 module.exports = router;
