+++ src/server.js
@@ -14,7 +14,7 @@ const usersRouter = require('./resources/users/users.router');
 const authRouter = require('./resources/auth/auth.router');
 const dislikeRouter = require('./resources/dislike/dislike.router');
 const likeRouter = require('./resources/like/like.router');
-
+const photoRouter = require ('./resources/photo/photo.router');
 //app
 dotenv.config();
 const app = express();
@@ -30,6 +30,7 @@ app.use('/api/dislike', dislikeRouter);
 app.use('/api/like', likeRouter);
 app.use('/api/users', usersRouter);
 app.use('/api/auth', authRouter);
+app.use('/api/photo', photoRouter);
 //token
 app.get(
   '/protected',
